+++ plugins/vitepress-plugin-permalink/build.config.ts
@@ -10,6 +10,6 @@ export default defineBuildConfig({
       exports: "named",
     },
   },
-  externals: ["vitepress", "vue"],
+  externals: ["vitepress", "vue", "vite"],
   failOnWarn: false,
 });