+++ vitepress-theme-teek/CHANGELOG.md
@@ -1,5 +1,17 @@
 # vitepress-theme-teek
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - vitepress-plugin-auto-frontmatter@1.0.6
+  - vitepress-plugin-catalogue@1.0.7
+  - vitepress-plugin-doc-analysis@1.0.8
+  - vitepress-plugin-file-content-loader@1.0.7
+  - vitepress-plugin-permalink@1.0.13
+  - vitepress-plugin-sidebar-resolve@1.0.6
+
 ## 1.0.11
 
 ### Patch Changes